Move config files in separate folder
