+++ scripts/validation/config/metadata-schema.json
