+++ scripts/validation/domain/validator.js
